Fully automated docker builds
This commit allows for fully automated Docker-based builds, using the
following scripts:

* `make_wheel.sh --docker-python${VERSION}` builds a wheel and automates
   the whole process within Docker including testing
* `test_wheel.sh` runs Python tests (optionally in Docker)
* `run_cpp_tests.py` now optionally runs C++ tests in Docker
* Travis CI now uses the above Docker config by default.
* Gitlab CI configuration provided; no server is hooked up. If you clone
  this repo to your own Gitlab repo, and provide runners, CI should work.

Fixes #257 -- with this change, we can now build release wheels directly
from this repo (thus picking up the correct git sha).

Fixes #1441 -- with this change, we are now compatible with libstdc++
from GCC 4.8 (compatible with Ubuntu 14.04).

+++ scripts/install_python_toolchain.sh
@@ -26,11 +26,13 @@ fi
 
 
 function linux_patch_sigfpe_handler {
-if [[ $OSTYPE == linux* ]]; then
-        targfile=deps/local/include/pyfpe.h
-        echo "#undef WANT_SIGFPE_HANDLER" | cat - $targfile > tmp
-        mv -f tmp $targfile
-fi
+  if [[ $OSTYPE == linux* ]]; then
+    targfile=deps/local/include/pyfpe.h
+    if [[ -f $targfile ]]; then
+      echo "#undef WANT_SIGFPE_HANDLER" | cat - $targfile > tmp
+      mv -f tmp $targfile
+    fi
+  fi
 }
 
 $PIP install --upgrade "pip>=8.1"
